requirements.txt
passlib==1.7.4
aio-pika==6.8.0
mongoengine==0.23.1
dnspython==2.1.0
yarl==1.5.1
celery==4.4.6
pyjwt==2.1.0
python-dateutil==2.8.2
pyfcm==1.5.3
